+++ jeap-deploymentlog-domain/src/main/java/ch/admin/bit/jeap/deploymentlog/domain/OpenFlowMetricValue.java
@@ -1,0 +1,4 @@
+package ch.admin.bit.jeap.deploymentlog.domain;
+
+public record OpenFlowMetricValue(String system, String component, FlowType type, long count) {
+}